refactor(resolver): remove duplicated dataloaders declarations
