+++ apps/backend/src/modules/shareable-resource/opencti/custom-view/custom-view.resolver.test.ts
@@ -4,46 +4,11 @@ import {
   contextSimpleUserFiligran2,
   GRAPHQL_RESOLVE_INFO,
 } from '../../../../../tests/tests.const';
-import {
-  CustomView,
-  Organization,
-  SubscriptionModel,
-} from '../../../../__generated__/resolvers-types';
-import ServiceInstance, {
-  ServiceInstanceId,
-} from '../../../../model/kanel/public/ServiceInstance';
-import UseCase from '../../../../model/kanel/public/UseCase';
-import User from '../../../../model/kanel/public/User';
+import { CustomView } from '../../../../__generated__/resolvers-types';
 import { DocumentChildrenDomain } from '../../../document/domain/document.children.domain';
-import { DocumentDomain } from '../../../document/domain/document.domain';
-import { ServiceInstanceDomain } from '../../../service/instance/service-instance.domain';
-import { subscriptionApp } from '../../../subscription/subscription.app';
-import { useCaseDomain } from '../../../use-case/use-case.domain';
 import customViewResolver from './custom-view.resolver';
 
 describe('customView field resolvers', () => {
-  describe('customView.use_cases', () => {
-    it('should load use cases by document id', async () => {
-      const documentId = uuidv4();
-      const expected = [{ id: uuidv4(), name: 'Use Case A' }];
-      vi.spyOn(useCaseDomain, 'loadUseCasesByDocumentId').mockResolvedValue(
-        expected as unknown as UseCase[]
-      );
-
-      const result = await customViewResolver.CustomView!.use_cases!(
-        { id: documentId } as unknown as CustomView,
-        {},
-        contextSimpleUserFiligran2,
-        GRAPHQL_RESOLVE_INFO
-      );
-
-      expect(useCaseDomain.loadUseCasesByDocumentId).toHaveBeenCalledWith(
-        documentId
-      );
-      expect(result).toEqual(expected);
-    });
-  });
-
   describe('customView.entity_types', () => {
     it('should parse entity types from the JSON metadata value', async () => {
       const expected = ['Attack-Pattern', 'Campaign'];
@@ -104,98 +69,6 @@ describe('customView field resolvers', () => {
       expect(
         DocumentChildrenDomain.loadImagesByDocumentId
       ).toHaveBeenCalledWith(documentId);
-      expect(result).toEqual(expected);
-    });
-  });
-
-  describe('customView.uploader', () => {
-    it('should load uploader by document id', async () => {
-      const documentId = uuidv4();
-      const expected = { id: uuidv4(), email: 'user@test.com' };
-      vi.spyOn(DocumentDomain, 'loadUploader').mockResolvedValue(
-        expected as unknown as User | undefined
-      );
-
-      const result = await customViewResolver.CustomView!.uploader!(
-        { id: documentId } as unknown as CustomView,
-        {},
-        contextSimpleUserFiligran2,
-        GRAPHQL_RESOLVE_INFO
-      );
-
-      expect(DocumentDomain.loadUploader).toHaveBeenCalledWith(documentId);
-      expect(result).toEqual(expected);
-    });
-  });
-
-  describe('customView.uploader_organization', () => {
-    it('should load uploader organization by document id', async () => {
-      const documentId = uuidv4();
-      const expected = { id: uuidv4(), name: 'Org A' };
-      vi.spyOn(DocumentDomain, 'loadUploaderOrganization').mockResolvedValue(
-        expected as unknown as Organization | undefined
-      );
-
-      const result = await customViewResolver.CustomView!
-        .uploader_organization!(
-        { id: documentId } as unknown as CustomView,
-        {},
-        contextSimpleUserFiligran2,
-        GRAPHQL_RESOLVE_INFO
-      );
-
-      expect(DocumentDomain.loadUploaderOrganization).toHaveBeenCalledWith(
-        documentId
-      );
-      expect(result).toEqual(expected);
-    });
-  });
-
-  describe('customView.service_instance', () => {
-    it('should load service instance by service_instance_id', async () => {
-      const serviceInstanceId = uuidv4() as ServiceInstanceId;
-      const expected = { id: serviceInstanceId, name: 'opencti custom views' };
-      vi.spyOn(ServiceInstanceDomain, 'getServiceInstance').mockResolvedValue(
-        expected as unknown as ServiceInstance | undefined
-      );
-
-      const result = await customViewResolver.CustomView!.service_instance!(
-        {
-          service_instance_id: serviceInstanceId,
-        } as unknown as CustomView,
-        {},
-        contextSimpleUserFiligran2,
-        GRAPHQL_RESOLVE_INFO
-      );
-
-      expect(ServiceInstanceDomain.getServiceInstance).toHaveBeenCalledWith(
-        serviceInstanceId
-      );
-      expect(result).toEqual(expected);
-    });
-  });
-
-  describe('customView.subscription', () => {
-    it('should load subscription model using context user and service_instance_id', async () => {
-      const serviceInstanceId = uuidv4() as ServiceInstanceId;
-      const expected = { id: uuidv4() } as unknown as SubscriptionModel;
-      vi.spyOn(subscriptionApp, 'loadSubscriptionModel').mockResolvedValue(
-        expected
-      );
-
-      const result = await customViewResolver.CustomView!.subscription!(
-        {
-          service_instance_id: serviceInstanceId,
-        } as unknown as CustomView,
-        {},
-        contextSimpleUserFiligran2,
-        GRAPHQL_RESOLVE_INFO
-      );
-
-      expect(subscriptionApp.loadSubscriptionModel).toHaveBeenCalledWith(
-        contextSimpleUserFiligran2.user,
-        serviceInstanceId
-      );
       expect(result).toEqual(expected);
     });
   });
